autonomous_sql_agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autonomous_sql_agent"
version = "0.1.0"
description = "Autonomous multi-agent SQL analysis pipeline for retail/e-commerce data warehouses"
requires-python = ">=3.12,<3.14"
license = {text = "MIT"}
dependencies = [
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
    "structlog==24.1.0",
    "rich==13.7.1",
    "pydantic-settings==2.2.1",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "transformers==4.41.2",
    "accelerate==0.30.1",
    "torch==2.3.0",
    "datasets==2.19.1",
    "sqlalchemy==2.0.30",
    "psycopg2-binary==2.9.9",
    "sqlglot==23.9.0",
    "plotly==5.22.0",
    "scikit-learn==1.4.2",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-cov==5.0.0",
    "ruff==0.4.4",
    "black==24.4.2",
    "mypy==1.10.0",
    "pre-commit==3.7.1",
    "pip-audit==2.7.3",
    "detect-secrets==1.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/autonomous_sql_agent"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "T20"]
ignore = ["S101"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
strict = true
python_version = "3.12"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/autonomous_sql_agent --cov-report=term-missing --cov-fail-under=60"

[tool.coverage.run]
source = ["src/autonomous_sql_agent"]
